requirements.txt
Flask
odoorpc
python-dotenv